requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pg8000>=1.30.0
openpyxl>=3.1.0